Former-commit-id: b83a0eec5dba4180d6b5dd8e7c7e8b3b8762c7f0

+++ user/aklokov/Mcrssemb.c
@@ -113,7 +113,7 @@ void readBlockAroundPoint (int xPos, int halfXapp, int* curXapp, int* leftShift)
 	memset (ptrToData_,   0, pointsNumToRead * sizeof (float));
 	memset (ptrToDataSq_, 0, pointsNumToRead * sizeof (float));			
 
-	const int startPos = startX * dagSize_ * sizeof (float);
+	const size_t startPos = (size_t) startX * dagSize_ * sizeof (float);
 
 	sf_seek (inDags_,   startPos, SEEK_SET);
 	sf_seek (inDagsSq_, startPos, SEEK_SET);
